Move error handling to a separate file
