+++ src/lib.rs
@@ -3,6 +3,7 @@ extern crate napi_derive;
 
 // Link other files
 pub mod auth;
+pub mod errors;
 pub mod options;
 pub mod paging;
 pub mod requests;
