+++ src/tests/utils_tests.rs
@@ -1,13 +1,14 @@
 use napi::bindgen_prelude::BigInt;
 
-use crate::utils::{self, bigint_to_i64};
+use crate::errors::js_error;
+use crate::utils::bigint_to_i64;
 
 #[napi]
 pub fn tests_bigint_to_i64(value: BigInt, case_id: Option<i32>) -> napi::Result<()> {
     let case_id = match case_id {
         Some(case_id) => case_id,
         None => {
-            return utils::bigint_to_i64(value, "Overflow expected").map(|_| ());
+            return bigint_to_i64(value, "Overflow expected").map(|_| ());
         }
     };
 
@@ -28,7 +29,7 @@ pub fn tests_bigint_to_i64(value: BigInt, case_id: Option<i32>) -> napi::Result<
             if v == expected {
                 Ok(())
             } else {
-                Err(utils::js_error(format!("Got {}, expected{}", v, expected)))
+                Err(js_error(format!("Got {}, expected{}", v, expected)))
             }
         }
         Err(e) => Err(e),
